Added LoRa communication support and updated GAS_ID
In preparation for integrating LoRa wireless communication, we've included the necessary hardware serial library and initiated setup for using a second hardware serial on the ESP32. The inclusion of LoRa communication support is aimed at extending the device's capability for remote data transmission, allowing for more flexible and longer-range connectivity options.

Furthermore, we've updated the Google App Script (GAS) deployment ID to ensure seamless data transfer to our backend. This change is critical for maintaining the integrity and functionality of our data pipeline.

As part of this update, we have also temporarily commented out sections related to the RTC and DHT sensors to focus on testing the new communication protocol. These sections will be re-enabled and integrated with the LoRa functionality once the initial tests are successful. To facilitate continuous operation and testing, fake sensor readings are temporarily used in the loop function.

This commit marks a pivotal step toward enhancing our device's communication capabilities, laying the groundwork for more advanced and reliable data transmission methods.

#IssueReference: If applicable